requirements.txt
amqp==5.1.0
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Flask==2.3.2
flower==1.2.0
gunicorn==20.1.0
humanize==4.0.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.29
pyparsing==3.0.8
pytz==2022.1
python-dotenv==1.0.0
redis==4.5.5
six==1.16.0
tornado==6.1
typing-extensions==4.1.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.3.4
wrapt==1.14.0
zipp==3.8.0
psycopg2-binary==2.9.6
Flask-SQLAlchemy==3.0.0
Flask-Mail==0.9.1
openai~=0.27.7
html2text~=2020.1.16
pytest~=7.3.1